Remove singleton variables.
That one branch of the left_taller and left_shorter functions, it only compares the feet and doesn't have to compare the inches.  We gave the inches names which we never used in any way, so SWI-Prolog gave us a warning about singleton variables.  Since we don't use the inches, let's not name them.